pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uniinfer"
version = "0.1.0"
description = "Hardware-agnostic AI inference runtime"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "UniInfer Contributors" },
]
keywords = ["inference", "llm", "ai", "hardware-agnostic", "llama-cpp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "llama-cpp-python>=0.3.0",
    "huggingface-hub>=0.20.0",
    "tokenizers>=0.15.0",
    "safetensors>=0.4.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
cuda = ["pynvml>=11.5.0"]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
uniinfer = "uniinfer.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/uniinfer"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
